continuing to improve based on checkstyle
